+++ compiler/src/Language/Mimsa/Modules/Check.hs
@@ -1,11 +1,8 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE GeneralisedNewtypeDeriving #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TupleSections #-}
 
-module Language.Mimsa.Modules.Check (checkModule, exprAndTypeFromParts) where
+module Language.Mimsa.Modules.Check (checkModule) where
 
 import Control.Monad.Except
 import Control.Monad.Reader
@@ -17,8 +14,9 @@ import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Helpers.Build as Build
+import Language.Mimsa.Modules.FromParts
 import Language.Mimsa.Modules.HashModule
-import Language.Mimsa.Modules.Prelude
+import Language.Mimsa.Modules.Monad
 import Language.Mimsa.Parser.Module
 import Language.Mimsa.Store
 import Language.Mimsa.Typechecker.DataTypes
@@ -33,11 +31,6 @@ import Language.Mimsa.Types.Modules.Module
 import Language.Mimsa.Types.Modules.ModuleHash
 import Language.Mimsa.Types.Store.ExprHash
 import Language.Mimsa.Types.Typechecker
-
--- this is where we keep all the modules we need to do things
-newtype CheckEnv ann = CheckEnv
-  { ceModules :: Map ModuleHash (Module ann)
-  }
 
 lookupModule :: ModuleHash -> CheckM (Module Annotation)
 lookupModule modHash = do
@@ -59,29 +52,18 @@ lookupModuleDep typecheckedModules name modHash = do
         _ -> throwError (ModuleErr (MissingModuleDep name modHash))
     _ -> throwError (ModuleErr (MissingModule modHash))
 
-newtype CheckM a = CheckM
-  { runCheckM ::
-      ExceptT
-        (Error Annotation)
-        ( Reader (CheckEnv Annotation)
-        )
-        a
-  }
-  deriving newtype
-    ( Functor,
-      Applicative,
-      Monad,
-      MonadError (Error Annotation),
-      MonadReader (CheckEnv Annotation)
-    )
-
-runCheck :: CheckM a -> Either (Error Annotation) a
-runCheck comp = runReader (runExceptT (runCheckM comp)) initialEnv
-  where
-    initialEnv =
-      CheckEnv
-        { ceModules = M.singleton preludeHash prelude
-        }
+lookupModuleType ::
+  Map ModuleHash (Module (Type Annotation)) ->
+  TypeName ->
+  ModuleHash ->
+  CheckM DataType
+lookupModuleType typecheckedModules typeName modHash = do
+  case M.lookup modHash typecheckedModules of
+    Just mod' ->
+      case M.lookup typeName (moDataTypes mod') of
+        Just dt -> pure dt
+        _ -> throwError (ModuleErr (MissingModuleTypeDep typeName modHash))
+    _ -> throwError (ModuleErr (MissingModule modHash))
 
 checkModule :: Text -> Either (Error Annotation) (Module (Type Annotation), MonoType)
 checkModule = runCheck . checkModule'
@@ -118,6 +100,8 @@ checkModule' input = do
   pure (tcMod, getModuleType tcMod)
 
 -- return type of module as a MTRecord of dep -> monotype
+-- TODO: module should probably be it's own MTModule or something
+-- as we'll want to pass them about at some point I think
 getModuleType :: Module (Type Annotation) -> Type Annotation
 getModuleType mod' =
   let defs =
@@ -189,9 +173,11 @@ typecheckAllModuleDeps typecheckedDeps inputModule = do
       <$> Build.doJobs (typecheckOneDep inputModule typecheckedDeps) state
   pure $ inputModule {moExpressions = typecheckedDefs}
 
-makeTypeDeclMap :: Module ann -> Map TyCon DataType
-makeTypeDeclMap inputModule =
-  M.fromList . fmap (first coerce) . M.toList $ moDataTypes inputModule
+makeTypeDeclMap :: Map TypeName DataType -> Module ann -> Map TyCon DataType
+makeTypeDeclMap importedTypes inputModule =
+  M.fromList . fmap (first coerce) . M.toList $
+    moDataTypes inputModule
+      <> importedTypes
 
 createTypecheckEnvironment ::
   Module Annotation ->
@@ -205,10 +191,15 @@ createTypecheckEnvironment inputModule deps typecheckedModules = do
       (lookupModuleDep typecheckedModules)
       (moExpressionImports inputModule)
 
+  importedTypes <-
+    M.traverseWithKey
+      (lookupModuleType typecheckedModules)
+      (moDataTypeImports inputModule)
+
   pure $
     createEnv
       (getTypeFromAnn <$> (deps <> importedDeps))
-      (makeTypeDeclMap inputModule)
+      (makeTypeDeclMap importedTypes inputModule)
 
 -- starting at a root module,
 -- create a map of each expr hash along with the modules it needs
@@ -288,139 +279,3 @@ typecheckOneDep inputModule typecheckedModules deps (name, expr) = do
         (ModuleErr . DefDoesNotTypeCheck name)
         (typecheck typeMap env numberedExpr)
   pure (first fst typedExpr)
-
-moduleFromModuleParts ::
-  (Monoid ann) =>
-  [ModuleItem ann] ->
-  CheckM (Module ann)
-moduleFromModuleParts parts =
-  let addPart part output = do
-        mod' <- output
-        case part of
-          ModuleExport modItem -> do
-            -- get whatever is inside
-            innerModule <- addPart modItem output
-            -- get the keys, add them to exports
-            pure $
-              innerModule
-                { moExpressionExports =
-                    M.keysSet (moExpressions innerModule)
-                }
-          ModuleExpression name bits expr ->
-            case M.lookup name (moExpressions mod') of
-              Just _ -> throwError (ModuleErr $ DuplicateDefinition name)
-              Nothing ->
-                let exp' = exprAndTypeFromParts bits expr
-                 in pure $
-                      mod'
-                        { moExpressions =
-                            M.singleton name exp' <> moExpressions mod'
-                        }
-          ModuleDataType dt@(DataType tyCon _ _) ->
-            let typeName = coerce tyCon
-             in case M.lookup typeName (moDataTypes mod') of
-                  Just _ -> throwError (ModuleErr $ DuplicateTypeName typeName)
-                  Nothing ->
-                    pure $
-                      mod'
-                        { moDataTypes =
-                            M.singleton typeName dt
-                              <> moDataTypes mod'
-                        }
-          ModuleImport (ImportAllFromHash mHash) -> do
-            importMod <- lookupModule mHash
-
-            let createImports =
-                  M.fromList
-                    . fmap (,mHash)
-                    . S.toList
-                    . moExpressionExports
-            pure $
-              mod'
-                { moExpressionImports =
-                    createImports importMod
-                      <> moExpressionImports mod'
-                }
-   in foldr addPart (pure mempty) parts
-
-addAnnotation :: Maybe (Type ann) -> Expr Name ann -> Expr Name ann
-addAnnotation mt expr =
-  -- add type annotation to expression
-  case mt of
-    Just typeAnnotation ->
-      MyAnnotation
-        (getAnnotationForType typeAnnotation)
-        typeAnnotation
-        expr
-    _ -> expr
-
--- given the bits of things, make a coherent type and expression
-exprAndTypeFromParts ::
-  (Monoid ann) =>
-  [DefPart ann] ->
-  Expr Name ann ->
-  Expr Name ann
-exprAndTypeFromParts parts expr =
-  let expr' =
-        foldr
-          ( \part rest -> case part of
-              (DefArg ident) -> MyLambda mempty ident rest
-              (DefTypedArg ident _) -> MyLambda mempty ident rest
-              (DefType _) -> rest
-          )
-          expr
-          parts
-      -- if we only have un-typed args, don't bother, we only want them as
-      -- placeholders
-      filteredParts =
-        let includesExplicitTypes =
-              any
-                ( \case
-                    (DefArg _) -> False
-                    _ -> True
-                )
-                parts
-            includesReturnType =
-              any
-                ( \case
-                    (DefType _) -> True
-                    _ -> False
-                )
-                parts
-         in if includesExplicitTypes
-              then
-                if includesReturnType
-                  then parts
-                  else parts <> [DefType (MTVar mempty (TVName "returnType"))]
-              else mempty
-      mt =
-        foldr
-          ( \part rest -> case part of
-              (DefArg (Identifier _ name)) -> case rest of
-                Just rest' ->
-                  Just
-                    ( MTFunction
-                        mempty
-                        (MTVar mempty (TVName (coerce name)))
-                        rest'
-                    )
-                Nothing ->
-                  Just
-                    ( MTVar
-                        mempty
-                        (TVName (coerce name))
-                    )
-              (DefTypedArg _ thisMt) -> case rest of
-                Just rest' ->
-                  Just
-                    (MTFunction mempty thisMt rest')
-                _ -> Just thisMt
-              (DefType thisMt) -> case rest of
-                Just rest' ->
-                  Just
-                    (MTFunction mempty rest' thisMt)
-                _ -> Just thisMt
-          )
-          Nothing
-          filteredParts
-   in addAnnotation mt expr'
