+++ compiler/src/Language/Mimsa/Modules/Prelude.hs
@@ -11,8 +11,10 @@ import qualified Data.Map as M
 import qualified Data.Set as S
 import Language.Mimsa.Modules.HashModule
 import Language.Mimsa.Types.AST
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.Modules.Module
 import Language.Mimsa.Types.Modules.ModuleHash
+import Language.Mimsa.Types.Typechecker
 
 preludeHash :: ModuleHash
 preludeHash = hashModule prelude
@@ -21,9 +23,11 @@ prelude :: Module Annotation
 prelude =
   Module
     { moExpressions = exprs,
-      moDataTypes = mempty,
+      moDataTypes = dts,
       moExpressionExports = S.singleton "fst",
-      moExpressionImports = mempty
+      moExpressionImports = mempty,
+      moDataTypeExports = S.singleton "Either",
+      moDataTypeImports = mempty
     }
   where
     exprs =
@@ -37,6 +41,19 @@ prelude =
                   (PPair mempty (PVar mempty "a") (PWildcard mempty))
                   (MyVar mempty "pair")
                   (MyVar mempty "a")
+              )
+          )
+        ]
+    dts =
+      M.fromList
+        [ ( "Either",
+            DataType
+              "Either"
+              ["e", "a"]
+              ( M.fromList
+                  [ ("Left", [MTVar mempty (TVName "e")]),
+                    ("Right", [MTVar mempty (TVName "a")])
+                  ]
               )
           )
         ]
