+++ compiler/test/modules/9.mimsa
@@ -1,0 +1,1 @@
+def id (str: String): a = str
